requirement.txt

Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.1
drf-yasg==1.21.7
psycopg2==2.9.9

